telegram_cover_bot v1/requirements.txt
requests==2.31.0
python-telegram-bot==20.6
Pillow==10.0.0
